requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
polib>=1.2.0
langchain==0.1.12
langchain-openai==0.0.8
